+++ app/Mail/Admin/CourseProposed.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Mail;
+namespace App\Mail\Admin;
 
 use App\Models\Course;
+use App\Models\User;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Mail\Mailable;
@@ -10,17 +11,22 @@ use Illuminate\Mail\Mailables\Content;
 use Illuminate\Mail\Mailables\Envelope;
 use Illuminate\Queue\SerializesModels;
 
-class CourseApproved extends Mailable
+class CourseProposed extends Mailable
 {
     use Queueable, SerializesModels;
 
     public $course;
+    public $teacher;
+    public $admin;
+
     /**
      * Create a new message instance.
      */
-    public function __construct(Course $course)
+    public function __construct(Course $course, User $admin)
     {
         $this->course = $course;
+        $this->teacher = $course->teacher;
+        $this->admin = $admin;
     }
 
     /**
@@ -29,7 +35,7 @@ class CourseApproved extends Mailable
     public function envelope(): Envelope
     {
         return new Envelope(
-            subject: '🟢 Tu curso ha sido aprobado',
+            subject: '📕 Nueva propuesta de curso',
         );
     }
 
@@ -39,8 +45,10 @@ class CourseApproved extends Mailable
     public function content(): Content
     {
         return new Content(
-            markdown: 'emails.course-approved',
-            with: ['show' => route('courses.show', $this->course)],
+            markdown: 'emails.admin.course-proposed',
+            with: [
+                'show' => route('admin.courses.show', $this->course),
+            ],
         );
     }
