+++ resources/views/emails/admin/course-proposed.blade.php
