+++ resources/views/emails/teacher/course-approved.blade.php
