+++ resources/views/emails/teacher/course-rejected.blade.php
